% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament_spec.R
\name{add_stage}
\alias{add_stage}
\title{Add a stage to a tournament specification}
\usage{
add_stage(spec, stage_id, stage)
}
\arguments{
\item{spec}{A \code{tournament_spec} object.}

\item{stage_id}{Unique stage identifier.}

\item{stage}{Stage definition object.}
}
\value{
Updated \code{tournament_spec}.
}
\description{
Add a stage to a tournament specification
}
