% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracket.R, R/double_elim.R, R/single_elim.R
\name{advance_winner}
\alias{advance_winner}
\alias{advance_winner.bracket}
\alias{advance_winner.double_elim_bracket}
\alias{advance_winner.single_elim_bracket}
\title{Advance winner to next match}
\usage{
advance_winner(bracket, match_id)

\method{advance_winner}{bracket}(bracket, match_id)

\method{advance_winner}{double_elim_bracket}(bracket, match_id)

\method{advance_winner}{single_elim_bracket}(bracket, match_id)
}
\arguments{
\item{bracket}{A bracket object}

\item{match_id}{The ID of the completed match}
}
\value{
Updated bracket object
}
\description{
Advance winner to next match
}
\keyword{internal}
