% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_seed_method}
\alias{apply_seed_method}
\title{Apply seed ordering to participants}
\usage{
apply_seed_method(participants_df, seed_method = "standard")
}
\arguments{
\item{participants_df}{data.frame from normalize_participants()}

\item{seed_method}{Character seed method}
}
\value{
data.frame reordered with a \code{seed_rank} column
}
\description{
Apply seed ordering to participants
}
\keyword{internal}
