% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_tiebreakers}
\alias{apply_tiebreakers}
\title{Apply tiebreaker ordering}
\usage{
apply_tiebreakers(df, matches, tiebreakers, allow_ties = TRUE)
}
\arguments{
\item{df}{Standings data frame with participant column}

\item{matches}{List of matches}

\item{tiebreakers}{Character vector of tiebreakers}

\item{allow_ties}{Logical}
}
\value{
Ordered data frame with rank column
}
\description{
Apply tiebreaker ordering
}
\keyword{internal}
