% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_groups}
\alias{assign_groups}
\title{Assign participants to groups using snake distribution}
\usage{
assign_groups(participants_df, groups)
}
\arguments{
\item{participants_df}{data.frame with participants}

\item{groups}{Number of groups}
}
\value{
data.frame with added \code{group} column
}
\description{
Assign participants to groups using snake distribution
}
\keyword{internal}
