% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament_rules.R
\name{bottom_n}
\alias{bottom_n}
\title{Select bottom ranked participants from source standings}
\usage{
bottom_n(n)
}
\arguments{
\item{n}{Positive integer count.}
}
\value{
A \code{bracketeer_selector} object.
}
\description{
Select bottom ranked participants from source standings
}
