% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament_rules.R
\name{bottom_per_group}
\alias{bottom_per_group}
\title{Select bottom ranked participants per group}
\usage{
bottom_per_group(n)
}
\arguments{
\item{n}{Positive integer count per group.}
}
\value{
A \code{bracketeer_selector} object.
}
\description{
Select bottom ranked participants per group
}
