% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracketeer-package.R
\docType{package}
\name{bracketeer-package}
\alias{bracketeer}
\alias{bracketeer-package}
\title{bracketeer: Tournament Generator}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Create and manage tournament brackets for various competition formats including single elimination, double elimination, round robin, Swiss system, and group-stage-to-knockout tournaments. Provides tools for seeding, scheduling, recording results, and tracking standings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bbtheo/bracketeer}
  \item \url{https://bbtheo.github.io/bracketeer/}
  \item Report bugs at \url{https://github.com/bbtheo/bracketeer/issues}
}

}
\author{
\strong{Maintainer}: Theo Blauberg \email{theo.blauberg@outlook.com} (\href{https://orcid.org/0000-0003-1780-0365}{ORCID}) [copyright holder]

}
\keyword{internal}
