% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_byes}
\alias{calc_byes}
\title{Calculate number of byes needed}
\usage{
calc_byes(n)
}
\arguments{
\item{n}{Number of participants}
}
\value{
Number of byes needed to reach next power of 2
}
\description{
Calculate number of byes needed
}
\keyword{internal}
