% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_rounds_single_elim}
\alias{calc_rounds_single_elim}
\title{Calculate number of rounds for single elimination}
\usage{
calc_rounds_single_elim(n)
}
\arguments{
\item{n}{Number of participants}
}
\value{
Number of rounds
}
\description{
Calculate number of rounds for single elimination
}
\keyword{internal}
