% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_match}
\alias{create_match}
\title{Create a match object}
\usage{
create_match(
  id,
  round,
  position,
  participant1 = NA_character_,
  participant2 = NA_character_,
  next_match = NA_integer_,
  next_slot = NA_integer_,
  bracket_type = "main",
  loser_next_match = NA_integer_,
  loser_next_slot = NA_integer_
)
}
\arguments{
\item{id}{Match ID}

\item{round}{Round number}

\item{position}{Position within round}

\item{participant1}{First participant (or NA for TBD)}

\item{participant2}{Second participant (or NA for TBD)}

\item{next_match}{ID of match winner advances to (or NA for final)}

\item{next_slot}{Which slot (1 or 2) in next match}

\item{bracket_type}{For double elim: "winners", "losers", or "grand_final"}
}
\value{
A bracket_match object
}
\description{
Create a match object
}
\keyword{internal}
