% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{default_tiebreakers}
\alias{default_tiebreakers}
\title{Default tiebreakers per format}
\usage{
default_tiebreakers(format)
}
\arguments{
\item{format}{Tournament format}
}
\value{
Character vector of tiebreakers
}
\description{
Default tiebreakers per format
}
\keyword{internal}
