% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament_export.R
\name{export_tournament_log}
\alias{export_tournament_log}
\title{Export tournament routing log entries}
\usage{
export_tournament_log(tournament)
}
\arguments{
\item{tournament}{A \code{tournament} object.}
}
\value{
Data frame with one row per routing log entry.
}
\description{
Export tournament routing log entries
}
