% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament_rules.R
\name{filter_by}
\alias{filter_by}
\title{Select participants using a custom predicate function}
\usage{
filter_by(fn)
}
\arguments{
\item{fn}{A transition predicate function.}
}
\value{
A \code{bracketeer_selector} object.
}
\description{
Select participants using a custom predicate function
}
