% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament_spec.R
\name{from_previous}
\alias{from_previous}
\title{Resolve source stage from most recently defined stage order}
\usage{
from_previous()
}
\value{
Sentinel object to be resolved by \code{add_transition()}.
}
\description{
Resolve source stage from most recently defined stage order
}
