% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament_runtime.R
\name{get_routing_log}
\alias{get_routing_log}
\title{Get transition routing log entries}
\usage{
get_routing_log(tournament)
}
\arguments{
\item{tournament}{A \code{tournament} object.}
}
\value{
List of routing log entries in append order.
}
\description{
Get transition routing log entries
}
