% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracket.R, R/double_elim.R,
%   R/group_stage_knockout.R, R/round_robin.R, R/single_elim.R, R/swiss.R
\name{get_standings}
\alias{get_standings}
\alias{get_standings.double_elim_bracket}
\alias{get_standings.group_stage_knockout}
\alias{get_standings.round_robin_bracket}
\alias{get_standings.single_elim_bracket}
\alias{get_standings.swiss_bracket}
\title{Get tournament standings}
\usage{
get_standings(bracket)

\method{get_standings}{double_elim_bracket}(bracket)

\method{get_standings}{group_stage_knockout}(bracket)

\method{get_standings}{round_robin_bracket}(bracket)

\method{get_standings}{single_elim_bracket}(bracket)

\method{get_standings}{swiss_bracket}(bracket)
}
\arguments{
\item{bracket}{A bracket object}
}
\value{
Data frame with standings
}
\description{
Get tournament standings
}
\details{
Standings are ordered using bracket-specific tiebreakers where applicable
(e.g., round robin and Swiss).
}
\keyword{internal}
