% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracket.R, R/double_elim.R,
%   R/group_stage_knockout.R, R/round_robin.R, R/single_elim.R, R/swiss.R,
%   R/tournament_runtime.R
\name{get_winner}
\alias{get_winner}
\alias{get_winner.bracket}
\alias{get_winner.double_elim_bracket}
\alias{get_winner.group_stage_knockout}
\alias{get_winner.round_robin_bracket}
\alias{get_winner.single_elim_bracket}
\alias{get_winner.swiss_bracket}
\alias{get_winner.tournament}
\title{Get tournament winner}
\usage{
get_winner(bracket)

\method{get_winner}{bracket}(bracket)

\method{get_winner}{double_elim_bracket}(bracket)

\method{get_winner}{group_stage_knockout}(bracket)

\method{get_winner}{round_robin_bracket}(bracket)

\method{get_winner}{single_elim_bracket}(bracket)

\method{get_winner}{swiss_bracket}(bracket)

\method{get_winner}{tournament}(bracket)
}
\arguments{
\item{bracket}{A bracket object}
}
\value{
Name of winner, or NA if tournament not complete
}
\description{
Get tournament winner
}
\keyword{internal}
