% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{head_to_head_points}
\alias{head_to_head_points}
\title{Compute head-to-head points within a subset}
\usage{
head_to_head_points(matches, participants, allow_ties = TRUE)
}
\arguments{
\item{matches}{List of matches}

\item{participants}{Character vector of participants}

\item{allow_ties}{Logical}
}
\value{
Named numeric vector of points
}
\description{
Compute head-to-head points within a subset
}
\keyword{internal}
