% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_bye_match}
\alias{is_bye_match}
\title{Check if a match is a bye (one participant is NA)}
\usage{
is_bye_match(match)
}
\arguments{
\item{match}{A bracket_match object}
}
\value{
Logical
}
\description{
Check if a match is a bye (one participant is NA)
}
\keyword{internal}
