% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracket.R, R/double_elim.R,
%   R/group_stage_knockout.R, R/swiss.R
\name{is_complete}
\alias{is_complete}
\alias{is_complete.bracket}
\alias{is_complete.double_elim_bracket}
\alias{is_complete.group_stage_knockout}
\alias{is_complete.swiss_bracket}
\title{Check if tournament is complete}
\usage{
is_complete(bracket)

\method{is_complete}{bracket}(bracket)

\method{is_complete}{double_elim_bracket}(bracket)

\method{is_complete}{group_stage_knockout}(bracket)

\method{is_complete}{swiss_bracket}(bracket)
}
\arguments{
\item{bracket}{A bracket object}
}
\value{
Logical
}
\description{
Check if tournament is complete
}
\keyword{internal}
