% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_power_of_2}
\alias{is_power_of_2}
\title{Check if n is a power of 2}
\usage{
is_power_of_2(n)
}
\arguments{
\item{n}{A positive integer}
}
\value{
Logical
}
\description{
Check if n is a power of 2
}
\keyword{internal}
