% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_elim.R
\name{new_double_elim_bracket}
\alias{new_double_elim_bracket}
\title{Internal double-elimination bracket constructor}
\usage{
new_double_elim_bracket(
  participants,
  seed = TRUE,
  grand_final_reset = TRUE,
  best_of = NULL,
  reseed = FALSE
)
}
\value{
A double_elim_bracket object.
}
\description{
Internal double-elimination bracket constructor
}
\keyword{internal}
