% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_robin.R
\name{new_round_robin_bracket}
\alias{new_round_robin_bracket}
\title{Internal round-robin bracket constructor}
\usage{
new_round_robin_bracket(
  participants,
  home_away = FALSE,
  n_rounds = NULL,
  best_of = NULL,
  tiebreakers = NULL,
  groups = NULL
)
}
\value{
A round_robin_bracket object.
}
\description{
Internal round-robin bracket constructor
}
\keyword{internal}
