% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_elim.R
\name{new_single_elim_bracket}
\alias{new_single_elim_bracket}
\title{Internal single-elimination bracket constructor}
\usage{
new_single_elim_bracket(
  participants,
  seed = TRUE,
  third_place = FALSE,
  best_of = NULL,
  reseed = FALSE
)
}
\value{
A single_elim_bracket object.
}
\description{
Internal single-elimination bracket constructor
}
\keyword{internal}
