% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_leg_knockout.R
\name{new_two_leg_bracket}
\alias{new_two_leg_bracket}
\title{Internal two-leg bracket constructor}
\usage{
new_two_leg_bracket(
  participants,
  seed = TRUE,
  third_place = FALSE,
  away_goals = TRUE,
  reseed = FALSE
)
}
\value{
A two_leg_knockout object.
}
\description{
Internal two-leg bracket constructor
}
\keyword{internal}
