% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{next_power_of_2}
\alias{next_power_of_2}
\title{Calculate the next power of 2}
\usage{
next_power_of_2(n)
}
\arguments{
\item{n}{A positive integer}
}
\value{
The smallest power of 2 >= n
}
\description{
Calculate the next power of 2
}
\keyword{internal}
