% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_seed_method}
\alias{normalize_seed_method}
\title{Normalize seed method}
\usage{
normalize_seed_method(seed)
}
\arguments{
\item{seed}{Logical or character}
}
\value{
Character seed method
}
\description{
Normalize seed method
}
\keyword{internal}
