% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_tiebreakers}
\alias{normalize_tiebreakers}
\title{Normalize tiebreaker configuration}
\usage{
normalize_tiebreakers(tiebreakers, format)
}
\arguments{
\item{tiebreakers}{Character vector or NULL}

\item{format}{Tournament format}
}
\value{
Normalized character vector
}
\description{
Normalize tiebreaker configuration
}
\keyword{internal}
