% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_stage_knockout.R, R/print.R
\name{print.group_stage_knockout}
\alias{print.group_stage_knockout}
\alias{print}
\alias{print.bracket_match}
\alias{print.bracket}
\alias{print.double_elim_bracket}
\alias{print.tournament}
\title{Print bracketeer objects}
\usage{
\method{print}{group_stage_knockout}(x, ...)

\method{print}{bracket_match}(x, ...)

\method{print}{bracket}(x, ...)

\method{print}{double_elim_bracket}(x, ...)

\method{print}{tournament}(x, ...)
}
\arguments{
\item{x}{A bracket or match object.}

\item{...}{Additional arguments (unused).}
}
\value{
The object, invisibly.
}
\description{
Print bracketeer objects
}
