% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament_inspection.R
\name{rankings}
\alias{rankings}
\title{Get tournament rankings}
\usage{
rankings(tournament)
}
\arguments{
\item{tournament}{A \code{tournament} object.}
}
\value{
Data frame of rankings.
}
\description{
Get tournament rankings
}
