% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament_rules.R
\name{remaining}
\alias{remaining}
\title{Select entrants remaining in the current transition source pool}
\usage{
remaining()
}
\value{
A \code{bracketeer_selector} object.
}
\description{
Select entrants remaining in the current transition source pool
}
