% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament_inspection.R
\name{routing_log}
\alias{routing_log}
\title{Get transition routing log}
\usage{
routing_log(tournament)
}
\arguments{
\item{tournament}{A \code{tournament} object.}
}
\value{
Data frame audit trail.
}
\description{
Get transition routing log
}
