% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament_spec.R
\name{set_outcome}
\alias{set_outcome}
\title{Configure tournament outcome depth}
\usage{
set_outcome(spec, track_placements = 1L)
}
\arguments{
\item{spec}{A \code{tournament_spec} object.}

\item{track_placements}{Number of placements to track.}
}
\value{
Updated \code{tournament_spec}.
}
\description{
Configure tournament outcome depth
}
