% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracket.R, R/group_stage_knockout.R
\name{set_winner}
\alias{set_winner}
\alias{set_winner.bracket}
\alias{set_winner.group_stage_knockout}
\title{Set match winner directly}
\usage{
set_winner(bracket, match_id, winner)

\method{set_winner}{bracket}(bracket, match_id, winner)

\method{set_winner}{group_stage_knockout}(bracket, match_id, winner)
}
\arguments{
\item{bracket}{A bracket object}

\item{match_id}{The ID of the match to update}

\item{winner}{Name of the winning participant}
}
\value{
Updated bracket object
}
\description{
Record the winner of a match without specifying scores.
}
\keyword{internal}
