% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament_rules.R
\name{slice_per_group}
\alias{slice_per_group}
\title{Select an inclusive standings slice per group}
\usage{
slice_per_group(from, to)
}
\arguments{
\item{from}{Positive integer starting position.}

\item{to}{Positive integer ending position (must be \verb{>= from}).}
}
\value{
A \code{bracketeer_selector} object.
}
\description{
Select an inclusive standings slice per group
}
