% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{snake_order}
\alias{snake_order}
\title{Generate a simple snake order}
\usage{
snake_order(n, block_size = 2L)
}
\arguments{
\item{n}{Number of items}

\item{block_size}{Block size to reverse in alternating fashion}
}
\value{
Integer vector of indices
}
\description{
Generate a simple snake order
}
\keyword{internal}
