% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament_inspection.R
\name{stage_status}
\alias{stage_status}
\title{Inspect tournament stage status}
\usage{
stage_status(tournament)
}
\arguments{
\item{tournament}{A \code{tournament} object.}
}
\value{
Data frame with one row per stage.
}
\description{
Inspect tournament stage status
}
