% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary}
\alias{summary}
\alias{summary.bracket}
\title{Summarize bracketeer objects}
\usage{
\method{summary}{bracket}(object, ...)
}
\arguments{
\item{object}{A bracket object.}

\item{...}{Additional arguments (unused).}
}
\value{
The object, invisibly.
}
\description{
Summarize bracketeer objects
}
