% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tiebreaker_chain}
\alias{tiebreaker_chain}
\title{Build an ordered tiebreaker chain}
\usage{
tiebreaker_chain(...)
}
\arguments{
\item{...}{Tiebreaker names. Accepts one or more strings or character
vectors.}
}
\value{
Character vector of unique tiebreakers in input order.
}
\description{
Build an ordered tiebreaker chain
}
\keyword{internal}
