% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tiebreaker_direction}
\alias{tiebreaker_direction}
\title{Determine tiebreaker direction}
\usage{
tiebreaker_direction(tiebreaker)
}
\arguments{
\item{tiebreaker}{Name}
}
\value{
"asc" or "desc"
}
\description{
Determine tiebreaker direction
}
\keyword{internal}
