% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament_rules.R
\name{top_n}
\alias{top_n}
\title{Select top ranked participants from source standings}
\usage{
top_n(n)
}
\arguments{
\item{n}{Positive integer count.}
}
\value{
A \code{bracketeer_selector} object.
}
\description{
Select top ranked participants from source standings
}
