% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament_rules.R
\name{top_per_group}
\alias{top_per_group}
\title{Select top ranked participants per group}
\usage{
top_per_group(n)
}
\arguments{
\item{n}{Positive integer count per group.}
}
\value{
A \code{bracketeer_selector} object.
}
\description{
Select top ranked participants per group
}
