% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament_spec.R
\name{tournament_spec}
\alias{tournament_spec}
\title{Create a tournament specification graph}
\usage{
tournament_spec()
}
\value{
A \code{tournament_spec} object.
}
\description{
Construct a multi-stage tournament specification object.
}
