% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_best_of}
\alias{validate_best_of}
\title{Validate best-of parameter}
\usage{
validate_best_of(best_of)
}
\arguments{
\item{best_of}{Integer or NULL}
}
\value{
Normalized integer or NULL
}
\description{
Validate best-of parameter
}
\keyword{internal}
