% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament_spec.R
\name{validate_tournament_spec}
\alias{validate_tournament_spec}
\title{Validate a tournament specification}
\usage{
validate_tournament_spec(spec)
}
\arguments{
\item{spec}{A \code{tournament_spec} object.}
}
\value{
The validated \code{tournament_spec}.
}
\description{
Validate a tournament specification
}
