% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament_inspection.R
\name{winner}
\alias{winner}
\title{Get tournament winner}
\usage{
winner(tournament)
}
\arguments{
\item{tournament}{A \code{tournament} object.}
}
\value{
Winner name or \code{NA_character_}.
}
\description{
Get tournament winner
}
\examples{
teams <- c("A", "B", "C", "D")
trn <- tournament(teams) |>
  round_robin("groups") |>
  single_elim("finals", take = top_n(2))

# ... enter all results ...

# Get the champion
winner(trn)
}
