% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracket.R, R/double_elim.R,
%   R/group_stage_knockout.R, R/single_elim.R, R/swiss.R, R/tournament_runtime.R
\name{advance}
\alias{advance}
\alias{advance.bracket}
\alias{advance.double_elim_bracket}
\alias{advance.group_stage_knockout}
\alias{advance.single_elim_bracket}
\alias{advance.swiss_bracket}
\alias{advance.tournament}
\title{Advance tournament to next round}
\usage{
advance(bracket, stage_id = NULL)

\method{advance}{bracket}(bracket, stage_id = NULL)

\method{advance}{double_elim_bracket}(bracket, stage_id = NULL)

\method{advance}{group_stage_knockout}(bracket, stage_id = NULL)

\method{advance}{single_elim_bracket}(bracket, stage_id = NULL)

\method{advance}{swiss_bracket}(bracket, stage_id = NULL)

\method{advance}{tournament}(bracket, stage_id = NULL)
}
\arguments{
\item{bracket}{A bracket object}

\item{stage_id}{Optional stage identifier for tournament methods.}
}
\value{
Updated bracket object
}
\description{
Check if current round is complete and update bracket state.
}
