% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swiss.R
\name{new_swiss_bracket}
\alias{new_swiss_bracket}
\title{Internal swiss bracket constructor}
\usage{
new_swiss_bracket(
  participants,
  rounds = NULL,
  seed = TRUE,
  allow_ties = TRUE,
  bye_points = 1,
  best_of = NULL,
  tiebreakers = NULL
)
}
\value{
A swiss_bracket object.
}
\description{
Internal swiss bracket constructor
}
\keyword{internal}
