% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bracket.R, R/tournament_runtime.R
\name{teardown}
\alias{teardown}
\alias{teardown.bracket}
\alias{teardown.tournament}
\title{Teardown tournament state}
\usage{
teardown(bracket, stage_id = NULL)

\method{teardown}{bracket}(bracket, stage_id = NULL)

\method{teardown}{tournament}(bracket, stage_id = NULL)
}
\arguments{
\item{bracket}{A bracket or tournament object.}

\item{stage_id}{Stage identifier to teardown (tournament method).}
}
\value{
Updated object.
}
\description{
For tournament runtimes, this un-materializes a stage and its downstream
dependents so upstream results can be corrected and replayed.
}
