% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament_inspection.R
\name{winner}
\alias{winner}
\title{Get tournament winner}
\usage{
winner(tournament)
}
\arguments{
\item{tournament}{A \code{tournament} object.}
}
\value{
Winner name or \code{NA_character_}.
}
\description{
Get tournament winner
}
